chat detail

+++ app/src/main/java/com/bitat/repository/sqlDB/SingleMsgDB.kt
@@ -65,7 +65,7 @@ object SingleMsgDB {
         selfId,
         otherId,
         time,
-        status,
+        status.toInt(),
         kind,
         content
     )
